refactor(coderd/templatebuilder): unexport ParseModulesFromFS, use internal test
- Unexport ParseModulesFromFS to parseModulesFromFS since it is only
  used internally and in tests.
- Move TestParseModulesFromFS into catalog_internal_test.go (package
  templatebuilder) so it can call the unexported function directly.
- Fix go:embed directive: remove invalid trailing slash from pattern.
